rev/rocky/src/requirements.txt
pycryptodome==3.22.0
